Did some plumbing for randomShuffle, changed the name from SI to splitCmd and added some todos
