+++ games/saveGame1.txt
@@ -1,0 +1,52 @@
+JS
+4C
+9C
+TD
+2D
+AH
+QD
+8H
+KD
+6S
+5C
+QC
+3D
+6D
+JC
+AD
+6H
+JH
+KC
+7C
+9H
+TH
+KS
+TS
+2H
+8C
+6C
+4D
+5D
+7D
+5S
+QH
+AC
+7H
+9S
+4H
+QS
+3S
+KH
+AS
+JD
+2S
+5H
+TC
+8D
+3C
+7S
+3H
+9D
+4S
+8S
+2C
